% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strength_bar.R
\name{updateStrengthBarInput}
\alias{updateStrengthBarInput}
\title{Actualizar strengthBar widget}
\usage{
updateStrengthBarInput(session, inputId, value, configuration = NULL)
}
\arguments{
\item{value}{Character. Valor nuevo que va a tomar el input}
}
\description{
Actualizar strengthBar widget
}
