% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strength_bar.R
\name{strengthBarInput}
\alias{strengthBarInput}
\title{React Password Strength Bar}
\usage{
strengthBarInput(
  inputId,
  label = "Contraseña",
  icon = "lock",
  scoreWords = c("Insegura", "Mala", "Regular", "Buena", "Excelente"),
  minLength = 6,
  shortScoreWord = "Muy Corta"
)
}
\arguments{
\item{inputId}{Character. Id del input.}

\item{label}{Character. Texto del input.}

\item{icon}{Character. Icono que aparece al lado del texto del input.}

\item{scoreWords}{Character. El nombre para cada estado de la barra, tienen que ser cinco.}

\item{minLength}{Numeric. La cantidad mínima de caracteres para que no aparezca la leyenda definida en \code{shortScoreWord}.}

\item{shortScoreWord}{Character. Texto que se muestra cuando no se alcanzan la cantidad mínima de caracteres.}
}
\value{
Lista con valor del input y nivel de seguridad.
}
\description{
Esta función proporciona un widget de entrada de contraseña para Shiny, que incluye una
barra de fortaleza de contraseña proveniente de la libreria
\href{https://www.npmjs.com/package/react-password-strength-bar}{react-password-strength-bar}.
La barra de fortaleza muestra la seguridad del texto
ingresado basándose en cinco posibles estados, siendo el estado mayor el más seguro.
El widget de entrada crea una lista en Shiny que contiene el texto ingresado y la
puntuación de seguridad, que varía de 0 a 4.
}
\examples{
\dontrun{
library(shiny)
devtools::load_all()

ui <- fluidPage(
  fontawesome::fa_html_dependency(),
  column(
    width = 6,
    strengthBarInput("textInput"),
    textOutput("textOutput"),
    actionButton("calcular", "Calcular")
  )
)

server <- function(input, output, session) {

  valor <- reactiveValues(password = "", score = 0)

  observeEvent(input$calcular, {
    valor$password <- input$textInput$password
    valor$score <- input$textInput$score
  })

  output$textOutput <- renderText({
    glue::glue(
      "Password: {valor$password} | Score: {valor$score}"
    )
  })
}

shinyApp(ui, server)
}
}
