% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strength_bar.R
\name{strengthBarInput}
\alias{strengthBarInput}
\title{React Password Strength Bar}
\usage{
strengthBarInput(
  inputId,
  default = list(password = ""),
  scoreWords = list("Insegura", "Mala", "Regular", "Buena", "Excelente"),
  minLength = 4,
  shortScoreWord = "Muy Corta"
)
}
\arguments{
\item{default}{Character. Valor que va a tomar el input por defecto}

\item{scoreWords}{List. El nombre para cada estado de la barra, tienen que ser cinco.}

\item{minLength}{Numeric. La cantidad minima de caracteres para que no aparezca la leyenda definida en \code{shortScoreWord}}

\item{shortScoreWord}{Character. Texto que se muestra cuando no se alcanzan la cantidad minima de caracteres}
}
\description{
Este input recibe una clave y e indica en la barra inferior y valida su fortaleza según cinco
estados posibles, siendo el mayor el de mayor seguridad.
}
