% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strength_bar.R
\name{strengthBarInput}
\alias{strengthBarInput}
\title{React Password Strength Bar}
\usage{
strengthBarInput(
  inputId,
  default = list(password = ""),
  scoreWords = c("Insegura", "Mala", "Regular", "Buena", "Excelente"),
  minLength = 4,
  shortScoreWord = "Muy Corta"
)
}
\arguments{
\item{inputId}{Character. Id del input.}

\item{default}{List. Valor que va a tomar el input por defecto, puede ser \code{NULL} también.}

\item{scoreWords}{Character. El nombre para cada estado de la barra, tienen que ser cinco.}

\item{minLength}{Numeric. La cantidad mínima de caracteres para que no aparezca la leyenda definida en \code{shortScoreWord}.}

\item{shortScoreWord}{Character. Texto que se muestra cuando no se alcanzan la cantidad mínima de caracteres.}
}
\value{
Un widget de entrada de shiny.
}
\description{
Esta función proporciona un widget de entrada de contraseña para Shiny, que incluye una
barra de fortaleza de contraseña proveniente de la libreria
\href{https://www.npmjs.com/package/react-password-strength-bar}{react-password-strength-bar}.
La barra de fortaleza muestra la seguridad del texto
ingresado basándose en cinco posibles estados, siendo el estado mayor el más seguro.
El widget de entrada crea una lista en Shiny que contiene el texto ingresado y la
puntuación de seguridad, que varía de 0 a 4.
}
\examples{
\dontrun{
library(shiny)

ui <- fluidPage(
  titlePanel("Ejemplo de Barra de Fortaleza de Contraseña"),
  sidebarLayout(
    sidebarPanel(
      strengthBarInput("contrasena", default = list(password = ""))
    ),
    mainPanel(
      verbatimTextOutput("output")
    )
  )
)

server <- function(input, output) {
  output$output <- renderPrint({
    input$contrasena
  })
}

shinyApp(ui = ui, server = server)
}
}
